# src/infra/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Documentação Swagger do prjeto de assinaturas",
        "version": "1.0.0",
        "title": "assinaturas",
        "contact": {
            "email": "keltonmof@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "schemes": [
        "http"
    ],
    "host": "189.48.177.66:3000",
    "basePath": "/",
    "paths": {
        "/usuarios/gravar-assinatura": {
            "post": {
                "summary": "Salva a assinatura de um usuário já cadastrado",
                "tags": ["Usuários"],
                "description": "Grava nos arquivos do computador a assinatura do usuário",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Objeto Resposta",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/postAssinaturaPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resposta criada",
                        "schema": {
                            "$ref": "#/definitions/postAssinaturaResponse"
                        }
                    },
                    "422": {
                        "description": "Parâmetro inválido",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        },
        "/usuarios": {
            "get": {
                "summary": "Retorna os usuários salvos",
                "tags": ["Usuários"],
                "description": "Lista os usuários previamente cadastrados e suas assinaturas se houver",
                "produces": [
                    "application/json"
                ],
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "query",
                        "name": "Nome",
                        "description": "Nome do usuário a ser pesquisado com Like",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "NomeCracha",
                        "description": "Nome no crachá do usuário a ser pesquisado com Like",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "CodigoUsuario",
                        "description": "Código do usuário pesquisado",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "limit",
                        "description": "Quantidade de registros na consulta",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "page",
                        "description": "Página com a quantidade de registros especificada.",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados recuperados com sucesso",
                        "schema": {
                            "$ref": "#/definitions/getUsuariosResponse"
                        }
                    },
                    "500": {
                        "description": "Erro no servidor de banco de dados",
                        "schema": {
                            "$ref": "#/definitions/InvalidResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "InvalidResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "postAssinaturaPayload": {
            "type": "object",
            "properties": {
                "CodigoUsuario": {
                    "type": "number"
                },
                "Assinatura": {
                    "type": "string",
                    "description": "Imagem em base64"
                }
            }
        },
        "postAssinaturaResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "CodigoUsuario": {
                    "type": "integer"
                },
                "CaminhoAssinatura": {
                    "type": "string"
                }
            }
        },
        "getUsuariosResponse": {
            "type": "object",
            "properties": {
                "statusCode": {
                    "type": "integer"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "CodigoUsuario": {
                                "type": "integer"
                            },
                            "Nome": {
                                "type": "string"
                            },
                            "NomeCracha": {
                                "type": "string"
                            },
                            "CaminhoAssinatura": {
                                "type": "string"
                            },
                            "Assinatura": {
                                "type": "string",
                                "description": "Assinatura em base64"
                            }
                        }
                    }
                },
                "metaData": {
                    "type": "object",
                    "properties": {
                        "totalItens": {
                            "type": "integer"
                        },
                        "totalPages": {
                            "type": "integer"
                        }
                    }
                }
            }
        }
    }
}